Made modifications to handle edge cases found
dowload_file() method changed to handle illegal directory names. If the
name of a directory to be created contains illegal characters this is now
crudely handled by storing the files downloaded in an invalid_path
directory in the related page directory.

gather_links() method has basic handling for web_link items found. All
web links for a specific page are written to a text file with the corresponding
name.

Changed the div element used that contains the courses material.