Moved saving web-links and added path validation
gather_links() now adds Web Links to the results output so that process_page()
can write it to file while processing the other results.

added fix_invalid_path() method which attempts to replace any illegal
characters from a path. When the path is still causing an error just use
the root course path to store files downloaded. This required adding
the course links to arguments for various methods to easily access the
save_path for each course when processing pages and downloading files.